Migrating to BigInt, same effect, even if postgresql supports bigint https://www.postgresql.org/docs/9.1/datatype-numeric.html, diesel does not seem to be ready

+++ migrations/2019-11-25-115919_create_htsget_blocks/up.sql
@@ -1,8 +1,8 @@
 CREATE TABLE htsget_blocks (
   bam_id VARCHAR PRIMARY KEY,
   target_name VARCHAR NOT NULL,
-  bytes_start INTEGER NOT NULL,
-  bytes_end INTEGER NOT NULL,
-  seq_start INTEGER NOT NULL,
-  seq_end INTEGER NOT NULL 
+  bytes_start BIGINT NOT NULL,
+  bytes_end BIGINT NOT NULL,
+  seq_start BIGINT NOT NULL,
+  seq_end BIGINT NOT NULL
 )
